Move settings header and items to separate components
